+++ packages/databricks-vscode/src/vscode-objs/CustomWhenContext.ts
@@ -85,16 +85,6 @@ export class CustomWhenContext {
         );
     }
 
-    updateShowWorkspaceView() {
-        commands.executeCommand(
-            "setContext",
-            "databricks.feature.views.workspace",
-            workspaceConfigs.experimetalFeatureOverides.includes(
-                "views.workspace"
-            )
-        );
-    }
-
     setIsActiveFileInActiveWorkspace(value: boolean) {
         commands.executeCommand(
             "setContext",
@@ -115,6 +105,14 @@ export class CustomWhenContext {
         commands.executeCommand(
             "setContext",
             "databricks.context.serverless",
+            value
+        );
+    }
+
+    setRemoteMode(value: boolean) {
+        commands.executeCommand(
+            "setContext",
+            "databricks.context.remoteMode",
             value
         );
     }